DROOLS-4522: Tests
