+++ drools-wb-screens/drools-wb-scenario-simulation-editor/drools-wb-scenario-simulation-editor-client/src/test/java/org/drools/workbench/screens/scenariosimulation/client/commands/ScenarioSimulationContextTest.java
@@ -71,5 +71,4 @@ public class ScenarioSimulationContextTest extends AbstractScenarioSimulationTes
         assertEquals(scenarioExpressionCellTextAreaSingletonDOMElementFactoryMock,
                      scenarioSimulationContextLocal.getScenarioExpressionCellTextAreaSingletonDOMElementFactory());
     }
-
 }
