+++ drools-wb-screens/drools-wb-scenario-simulation-editor/drools-wb-scenario-simulation-editor-client/src/test/java/org/drools/workbench/screens/scenariosimulation/client/utils/ScenarioSimulationUtilsTest.java
@@ -117,16 +117,4 @@ public class ScenarioSimulationUtilsTest extends AbstractUtilsTest {
         retrieved = ScenarioSimulationUtils.getPropertyNameElementsWithoutAlias(Arrays.asList(aliasName, propertyName), factIdentifierMock);
         assertEquals(Arrays.asList(className, propertyName), retrieved);
     }
-
-    /*@Test
-    public void getPlaceholder() {
-        assertEquals(ScenarioSimulationEditorConstants.INSTANCE.insertValue(),
-                     ScenarioSimulationUtils.getPlaceholder(String.class.getCanonicalName()));
-
-        assertEquals(ScenarioSimulationEditorConstants.INSTANCE.dateFormatPlaceholder(),
-                     ScenarioSimulationUtils.getPlaceholder(LocalDate.class.getCanonicalName()));
-
-        assertEquals(ScenarioSimulationEditorConstants.INSTANCE.dmnDateFormatPlaceholder(),
-                     ScenarioSimulationUtils.getPlaceholder(ConstantHolder.DMN_DATE));
-    }*/
 }
